requirements.txt
absl-py==0.9.0
astor==0.8.1
backcall==0.2.0
cycler==0.10.0
decorator==4.4.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
imageio==2.8.0
imgaug==0.4.0
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
Keras==2.0.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.2.1
matplotlib==3.2.1
networkx==2.4
numpy==1.18.2
opencv-python==4.2.0.34
opt-einsum==3.2.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
pycocotools==2.0.0
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-image==0.16.2
scipy==1.4.1
Shapely==1.7.0
six==1.14.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.2
termcolor==1.1.0
Theano==1.0.4
traitlets==4.3.3
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
